Speedup adding clauses by not sorting unary and binary clauses
This works because the reason MiniSat sorts the clause is to run
a unique-like loop on it. Unary clause is sorted by definition,
and binary clauses are "sorted enough" for unique.

This change provides 2-3 % speedup for setting up Sat instance in
the internal CyberCalc.Golem project.

Caveats for the measurement:
* CyberCalc.Golem is _very_ heavy on binary clauses (>50% of all
  generated clauses are binary).
* CyberCalc.Golem touches lot of other stuff while setting up the
  Sat instance, so the likely speedup for specifically adding
  should be larger.
* Measurement was done as a macro benchmark in a specific LTOd binary